requirements.txt
absl-py==2.1.0
accelerate==0.29.3
astunparse==1.6.3
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
diffusers==0.27.2
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.3.1
functorch  
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
huggingface-hub==0.22.2
idna==3.7
importlib-metadata==7.1.0
intel-openmp==2021.4.0 
itsdangerous==2.2.0
Jinja2==3.1.3
keras==2.15.0
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
mkl==2021.4.0
ml_dtypes==0.2.0
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.3.0
packaging==24.0
Pillow==10.3.0
pip
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
requests_oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
setuptools==65.5.0
six==1.16.0
sympy==1.12
tbb==2021.12.0
urllib3==2.2.1
werkzeug==3.0.2
wheel==0.43.0
wrapt==1.14.1
zipp==3.18.1 